# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OmniDetect API",
    "description": "An enterprise-grade, high-performance, stateless API combining linguistic perplexity AI content detection, active-voice/rhythm bypass humanization, and search index plagiarism validation.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://omnidetect-api.onrender.com",
      "description": "Production Render Server"
    },
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/v1/detect": {
      "post": {
        "summary": "AI Content Detection",
        "description": "Analyzes text character layout, Shannon word entropy, lexical richness indices, N-gram patterns, and sentence length standard deviation to calculate probability signatures of AI-generated content. Supports optional LLM-assisted style co-scans.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Text body content to check (up to 50,000 characters)."
                  },
                  "hybrid": {
                    "type": "boolean",
                    "description": "If true, combines statistical heuristics with deep learning styling analysis (requires server GEMINI_API_KEY / OPENAI_API_KEY).",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Linguistic Heuristics AI probability score results.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "aiProbability": 87,
                  "label": "AI",
                  "language": "en",
                  "classifiers": {
                    "gptProbability": 45,
                    "claudeProbability": 25,
                    "geminiProbability": 12,
                    "llamaProbability": 5
                  },
                  "metrics": {
                    "wordCount": 128,
                    "sentenceCount": 8,
                    "lexicalDiversity": 0.54,
                    "giraudIndex": 6.11,
                    "herdanIndex": 0.88,
                    "entropy": 5.12,
                    "burstiness": 1.25,
                    "predictabilityScore": 82.4,
                    "aiKeywordDensity": 0.08,
                    "bigramRepetitionRatio": 0.95,
                    "trigramRepetitionRatio": 0.98
                  },
                  "sentences": [
                    {
                      "text": "Furthermore, it is important to remember that tapestries of technology foster synergy.",
                      "wordCount": 12,
                      "aiScore": 95,
                      "highlightRisk": "high"
                    }
                  ],
                  "hybridScanPerformed": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/humanize": {
      "post": {
        "summary": "AI Bypass Humanizer",
        "description": "Rewrites predictable transitions, randomizes sentence rhythm layouts, and inserts conversational contractions to bypass AI probability scans. Supports programmatic rewriting and high-fidelity LLM-assisted rephrasing.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "AI-generated text content to rewrite (up to 10,000 characters)."
                  },
                  "mode": {
                    "type": "string",
                    "enum": ["standard", "creative", "academic", "premium"],
                    "default": "standard",
                    "description": "Rewriting strength and syntax mutation density. Premium mode utilizes LLM-based rephrasing."
                  },
                  "readability": {
                    "type": "string",
                    "enum": ["elementary", "high_school", "college", "professional"],
                    "description": "Target readability grade level (Premium mode only)."
                  },
                  "tone": {
                    "type": "string",
                    "enum": ["casual", "academic", "creative", "technical"],
                    "description": "Target tone of voice style (Premium mode only)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Humanized rewritten text and verification score.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "originalText": "Furthermore, it is important to remember that tapestries of technology foster synergy.",
                  "humanizedText": "Also, keep in mind that complex collections of technology build teamwork.",
                  "originalScore": 87,
                  "newScore": 8,
                  "wordCount": 11,
                  "readability": {
                    "before": {
                      "readingEase": 45.5,
                      "gradeLevel": "High School"
                    },
                    "after": {
                      "readingEase": 75.2,
                      "gradeLevel": "Middle School"
                    }
                  },
                  "humanizeMethod": "programmatic"
                }
              }
            }
          }
        }
      }
    },
    "/v1/check/plagiarism": {
      "post": {
        "summary": "Plagiarism Checker",
        "description": "Queries web index registries and Wikipedia fact databases to compare sentence overlap percentages and flag duplicate copy. Employs blended Levenshtein-distance and Jaccard token metrics.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Text to analyze for plagiarism matches."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Web index similarity validation details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "plagiarismPercent": 20,
                  "uniquePercent": 80,
                  "matches": [
                    {
                      "matchedText": "Cat is a small domesticated carnivorous mammal.",
                      "sourceUrl": "https://en.wikipedia.org/wiki/Cat",
                      "sourceTitle": "Cat - Wikipedia",
                      "similarity": 95.5,
                      "startIndex": 0,
                      "endIndex": 46
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}